requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.4.0
pycparser==2.20
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
xml-python==0.4.3
xmltodict==0.12.0
zipfile36==0.1.3
